backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.0
google-cloud-storage==2.14.0
google-generativeai==0.3.2
redis==5.0.1
pydantic==2.5.3
python-multipart==0.0.6
websockets==12.0
aiohttp==3.9.1
